+++ src/Loading.php
@@ -247,12 +247,12 @@ class Loading {
 			foreach ( $meta['sizes'] as $size ) {
 				// check if width matches width source and if ratio is within 1px.
 				if ( $size['width'] === $source['value'] &&
-				     wp_image_matches_ratio(
-					     (int) $size['width'],
-					     (int) $size['height'],
-					     (int) $sizes[0],
-					     (int) $sizes[1]
-				     )
+					wp_image_matches_ratio(
+						(int) $size['width'],
+						(int) $size['height'],
+						(int) $sizes[0],
+						(int) $sizes[1]
+					)
 				) {
 					$source['url'] = $this->replace_url( $this->get_attachment_url( $source['url'], $attachment_id ), $size );
 					continue 2;